requirements.txt
# ÖNIKA LI - FastAPI版本
fastapi==0.109.0
uvicorn==0.27.0
python-telegram-bot==20.7
openai>=1.0.0
anthropic>=0.18.0
python-multipart==0.0.6
